+++ Core/Nanjing.h
@@ -20,8 +20,8 @@ private:
 	}
 
 protected:
-	virtual uint32_t GetPRGPageSize() { return 0x8000; }
-	virtual uint32_t GetCHRPageSize() {	return 0x1000; }
+	virtual uint16_t GetPRGPageSize() { return 0x8000; }
+	virtual uint16_t GetCHRPageSize() {	return 0x1000; }
 
 	virtual void StreamState(bool saving)
 	{
